refactor(web): extract TaskPayloadBuilder from poll route
Extract ~130 lines of task payload assembly logic from the poll route
into a reusable TaskPayloadBuilder service class. This enables:
1. Cleaner separation of concerns
2. Reusability for WebSocket push logic (Task 6)
3. Easier testing and maintenance

The poll route now delegates to TaskPayloadBuilder and patches the
user_email field (which is only available in the poll context via auth).

All 992 tests pass. Behavior is identical.
